+++ packages/tools/utils/src/array/fn/object/array.js
@@ -5,6 +5,8 @@
 import * as _ from "lamb";
 import {__} from "lamb";
 
+import {concat} from "../../proto/array.js";
+
 /**
  * Return a function expecting an object and returning an array of values corresponding to the provided keys.
  *
@@ -28,3 +30,29 @@ export const makeKeysGetter = _.pipe([
     _.mapWith(_.getKey),
     _.collect
 ]);
+
+/**
+ * Return a function expecting an object and concatenating values in the provided whitelisted keys.
+ *
+ * @function
+ * @arg {array} keys - Array of keys
+ * @return {function} - Object -> Array
+ *
+ * @example
+const getProducts = pickAndConcatValues(['food', 'beverage']);
+getProducts({
+    food: ['bread', 'cheese', 'ham'],
+    beverage: ['wine', 'water'],
+    id: 'area1',
+    value: 32.1,
+});
+// ['bread', 'cheese', 'ham', 'wine', 'water']
+ *
+ * @version 0.4.0
+ */
+export const pickAndConcatValues =
+  whitelist => _.pipe([
+    _.pickKeys(whitelist),
+    _.values,
+    _.apply(concat)
+  ]);