+++ CHANGELOG.md
@@ -17,6 +17,10 @@
 
 - added `isNotNull`
 
+### Array -> (Object -> Array)
+
+- added `pickAndConcatValues`
+
 ### Array -> (String -> Boolean)
 
 - added `containsOneOf`